requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
fpdf2>=2.7.0
ultralytics>=8.0.0
PyYAML>=6.0

